requirements.txt
python-telegram-bot[job-queue]==20.7
anthropic==0.17.0
python-dotenv==1.0.1
requests==2.31.0
openai